[MODULARIZATION] Refactor application structure and update dependencies for multi-module support
