+++ eshop/.run/Run eShop (PostgreSQL).run.xml
@@ -3,7 +3,7 @@
     <option name="ACTIVE_PROFILES" value="pg" />
     <option name="FRAME_DEACTIVATION_UPDATE_POLICY" value="UpdateClassesAndResources" />
     <module name="main" />
-    <option name="SPRING_BOOT_MAIN_CLASS" value="lt.codeacademy.spring2025.EshopApplication" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="lt.codeacademy.spring2025.eshop.EshopApplication" />
     <option name="UPDATE_ACTION_UPDATE_POLICY" value="UpdateClassesAndResources" />
     <extension name="net.ashald.envfile">
       <option name="IS_ENABLED" value="false" />
