+++ eshop/security/basic-mvc/src/main/java/lt/codeacademy/security/config/BasicSecurityConfig.java
@@ -12,8 +12,6 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.util.matcher.AntPathRequestMatcher;
 
-import javax.sql.DataSource;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import lt.codeacademy.security.service.UserService;
@@ -24,7 +22,6 @@ import lt.codeacademy.security.service.UserService;
 @EnableMethodSecurity(prePostEnabled = true, securedEnabled = true, jsr250Enabled = true)
 public class BasicSecurityConfig {
 
-  private final DataSource datasource;
   private final UserService userService;
 
   @Bean
